Housekeeping
The api_error function now sets the Content-Type header to
application/json.

The api_login.php controller now properly passes the "real_ip" through
when resuming a session.

The static function on the User class which allows for resuming a
session now takes a $real_ip parameter, which it should've all along.

Completed the Resume static function (which resumes sessions) so that it
will now correctly update per session data in the database.

Pulled assets from the MISA devel site and rehosted them (since the MISA
devel site requires authentication to access, demoing the site was
painful as there was two sets of logins to deal with).

Added Ontario to the province/state/country drop-down.  Apparently it
was overlooked.

Bcrypt slowness issues were fixed by going from 15 rounds to 12.  Code
that was previously commented out in an attempt to localize the problem
has been restored (as it was not the problem, the number of rounds was).

Created an organizations.php file, as an organizations administration
page is the next task on the to do list.