Organization Form
Added organization form.

Fixed bug with phone/fax validation on the user form.

Updated the Organization class to support containing database rows or
associative arrays.

Added the ability to get the name of an organization's membership type
from the Organization object.

Added the ability to save an Organization back to the database.

Fixed a bug in the debug error handling in main.php which was preventing
error messages from being printed in debug mode.

Fixed a small display issue with display_user.phtml.