Listings, Bugfixes, Cosmetic Polish, API Changes
Added the ability to query information about organizations to the API.
Pages of organizations, the number of organizations, and specific
organizations may all be queried.

API calls for organization and user counts, and for pages of
organizations and users may be filtered based on the status of the
users/organizations.

All pages have appropriate titles.

There's now an "Organizations" section where a paginated listing of
organizations is available.

There's now a "Users" section where a paginated listing of users is
available.

Added an indicator that a user is logged in to the upper-right corner of
the UI.  This also communicates who they're logged in as.

Updated the schema.

Added the ability for users to be logged in/have a session cookie set
from a session key passed on the query string.  This allows users logged
in through API calls to be seamlessly directed to pages on this site
without having to login again.

Updated the schema change script so that user types are scraped, so that
country/province/state is populated correctly, and so that payment
information is scraped.

Fixed a bug in the Request class.