API Update
Fixed a bug that was preventing API requests from being property routed.

Incorporated query API changes into the JSON API.  Queries for users and
organizations based on whether they are active or inactive, or based on
a keyword search is now supported.  Queries may be paginated or not.
Organizations may be queried based on whether they are pending or not.