Password Reset
Updated the schema to make `activation_key` the proper size (32
hexadecimal digits).

Implemented password reset functionality.

Added an EMail class to simplify the sending of templated HTML e-mails.

Fixed a bug in the login template which was causing query string
arguments not to be properly forwarded when a user clicked a link with
query string arguments, but had to login before proceeding.