Add Organizations
Added the ability to add new organizations without going through the
membership application process.  This is exposed through the
"Administration" UI, since it is an admin-only function.

Fixed a bug with table locking in the membership application form.

Added a missing null check to the person controller.

Updated the organizations' users and organizations' payment info
templates so that they display a message rather than an empty table when
there are no results.