Add Users (UI)
Added links to the add users functionality throughout the UI.  Sitewide
admins get a link on the admin page and on every organization's page,
organization admins get it on their organization's page.

A few bugfixes:

Revamped the way login is handled, a 3xx is now sent to get rid of the
login=true query string, which eliminates some bug prone handling
related to that.

Added a null check to the person controller so that superusers with no
organization can't create more users.