Top Menu Hook and Remember Me
Added ability to toggle long-lived cookies on/off at the login screen.
If this is toggled off the system will send cookies which will be
cleared when the user's browser is closed.

Added a hook for fetching the top menu as JSON from the MISA WordPress
site.  This has not actually been hooked in -- i.e. the new top menu is
not actually displayed -- but the functionality exists and has been
tested.