Login API Exposes Unpaid Duration
The login API will now set a property called "duration" on its response
when a login attempt returns that the user's organization's dues are
unpaid.

This allows the API consumer to make a decision about whether the user
has been unpaid for a short enough duration to be allowed to login under
some grace period.