Schema Update
Updated the schema to support new payment information.

The membership dues-related fields (IDs which link payments to
membership years and types) are now nullable so that other payments may
be stored in that table.

There is now a "response" field which will be used to store the raw
credit card receipt from e-xact.