Edit/Add Person -- Username to E-Mail Switch
Made changes/fixes to person.php -- the controller responsible for
allowing editing and addition of users.

Fixed a bug that was causing the "add" argument in the URL not to be
picked up, so the controller was forever stuck in edit mode and would
not allow new users to be created.

Switched over from usernames and e-mails to just e-mails.  Usernames are
no longer displayed, prompted for, or checked.

Added consistent checking for duplicate e-mails on postback, and added
graceful handling when a duplicate e-mail is detected -- an error
message is displayed and the user's data is retained in the form.