Receipt Links
Organization pages now have a link to receipt pages for each paid
payment.

Just as with all payment details, these links are visible only to admins
and superusers associated with that organization.

In the future -- when the ability to edit payments is added --
administrators' link will link to the payment edit page.