Add/Edit Payment Links
On the organization page, superusers of that organization will be able
to click on each payment (which is paid) to go to the receipt for that
payment.

Administrators, however, will be able to click any payment to be taken
to the edit payment page for that payment (from which they can click
through to view the receipt if the payment is flagged as having been
paid).