Delete Users
Users can now be deleted from their "Edit User" page.

A user may delete another user if they are not trying to delete
themselves and they are an administrator, or they are a superuser of the
organization the target user is a member of.

This is laid out in detail in person.php lines 47-63.