Search Overhaul
The search in the top-right is now always present, and uses the MISA BC WordPress site's search.

Added a "Search" link to the sidebar for admins only, which takes you to the "keyword" controller.

The "keyword" controller now has a front-end, which is simply a form which allows you to enter search terms.