GET Session Key Redirect
The system allows users to "login" by visiting the site with a session key in the query string.  This enables seamless transition of sessions from the WordPress site to the membership management system.

However, it is desirable -- once the cookie has been set -- to remove this sensitive information from the browser.

Accordingly, the system now redirects to the same URL minus the query string after setting the cookie.