Password Reset E-Mail Fix
Fixed an issue with password reset e-mails: When the system was changed from usernames to e-mails, the password reset e-mails was not updated.  Therefore the e-mails the system sent did not properly allow the system to look up the user whose password was to be changed, causing the entire mechanism to fail.